requirements.txt
djangorestframework
Django
cryptography
psycopg2
